requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
CacheControl==0.14.3
cachetools==5.5.2
celery==5.3.0
certifi==2025.4.26
cffi==1.17.1
channels==4.0.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==42.0.0
dj-database-url==2.1.0
Django==4.2
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.0
firebase-admin==6.8.0
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
msgpack==1.1.0
mysqlclient==2.2.7
pillow==11.2.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
rsa==4.9.1
setuptools==80.8.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
channels-redis
gunicorn==21.2.0
daphne>=4.0
